(BSR)[API] chore: move last dms script into core
