--- api/src/pcapi/scripts/subscription/dms/__init__.py
@@ -1,1 +1,0 @@
-from .handle_inactive_dms_applications import handle_inactive_dms_applications
